+++ back/account/src/main/java/com/bantads/account/account/repository/AccountRepository.java
@@ -1,8 +1,9 @@
-package com.bantads.account.repository;
+package com.bantads.account.account.repository;
 
-import com.bantads.account.model.Account;
 import org.springframework.data.jpa.repository.*;
 import org.springframework.stereotype.Repository;
+
+import com.bantads.account.account.models.Account;
 
 @Repository
 public interface AccountRepository extends JpaRepository<Account, Long> {
