+++ back/account/src/main/java/com/bantads/account/transaction/models/Transaction.java
@@ -1,27 +1,48 @@
-package com.bantads.account.model;
+package com.bantads.account.transaction.models;
 
 import java.io.Serializable;
 
+import javax.persistence.*;
+
 import org.modelmapper.ModelMapper;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonRawValue;
+import com.bantads.account.account.models.Account;
 
-@JsonInclude(JsonInclude.Include.NON_EMPTY)
-public class TransactionDTO implements Serializable {
+@Entity
+@Table(name = "transactions")
+public class Transaction implements Serializable {
+    private static final Long serialVersionUID = 1L;
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
     private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Account account;
+
+    @Column(name = "type")
     private String type;
+
+    @Column(name = "amount")
     private Double amount;
+
+    @Column(name = "timestamp")
     private Long timestamp;
-    @JsonRawValue
+
+    @Column(name = "extra_data")
     private String extraData;
+
+    @Column(name = "balance_before")
     private Double balanceBefore;
 
-    public TransactionDTO() {
+    public Transaction() {
     }
 
-    public TransactionDTO(Long id, String type, Double amount, Long timestamp, String extraData, Double balanceBefore) {
+    public Transaction(Long id, Account account, String type, Double amount,
+            Long timestamp, String extraData, Double balanceBefore) {
         this.id = id;
+        this.account = account;
         this.type = type;
         this.amount = amount;
         this.timestamp = timestamp;
@@ -29,9 +50,9 @@ public class TransactionDTO implements Serializable {
         this.balanceBefore = balanceBefore;
     }
 
-    public Transaction toEntity() {
+    public TransactionDTO toDto() {
         ModelMapper mapper = new ModelMapper();
-        return mapper.map(this, Transaction.class);
+        return mapper.map(this, TransactionDTO.class);
     }
 
     public Long getId() {
@@ -72,6 +93,18 @@ public class TransactionDTO implements Serializable {
 
     public void setExtraData(String extraData) {
         this.extraData = extraData;
+    }
+
+    public static Long getSerialversionuid() {
+        return serialVersionUID;
+    }
+
+    public Account getAccount() {
+        return account;
+    }
+
+    public void setAccount(Account account) {
+        this.account = account;
     }
 
     public Double getBalanceBefore() {
