+++ back/account/src/main/java/com/bantads/account/transaction/rest/TransactionQueries.java
@@ -1,4 +1,4 @@
-package com.bantads.account.rest.transaction;
+package com.bantads.account.transaction.rest;
 
 import java.util.ArrayList;
 
@@ -9,13 +9,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.bantads.account.account.models.Account;
+import com.bantads.account.account.services.AccountServices;
 import com.bantads.account.exceptions.AccountNotFound;
 import com.bantads.account.lib.JsonResponse;
 import com.bantads.account.lib.Statement;
-import com.bantads.account.model.Account;
-import com.bantads.account.model.TransactionDTO;
-import com.bantads.account.services.AccountServices;
-import com.bantads.account.services.TransactionServices;
+import com.bantads.account.transaction.models.TransactionDTO;
+import com.bantads.account.transaction.services.TransactionServices;
 
 @CrossOrigin
 @RestController
